requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.8
kafka-python>=2.0.2
psycopg2-binary~=2.9.10
SQLAlchemy~=2.0.38